pyproject.toml
[project]
name = "self-talking-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
    "ipython>=8.0.0",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
]
